requirements.txt
streamlit
groq
python-docx
python-pptx
pandas
fpdf
python-dotenv
watchdog
